app/requirements.txt
Django==3.0.7
djangorestframework==3.11.1
psycopg2-binary==2.8.4
